# apps/api/src/routes/openapi-v1.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "PhotoBrain Native API",
		"version": "1.0.0"
	},
	"paths": {
		"/api/v1/folders": {
			"get": {
				"operationId": "listFolders",
				"responses": {
					"200": {
						"description": "Folder tree",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FoldersResponse"
								}
							}
						}
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		},
		"/api/v1/filter-options": {
			"get": {
				"operationId": "listFilterOptions",
				"parameters": [
					{
						"name": "folder",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Available filters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FilterOptionsResponse"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/InvalidRequest"
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		},
		"/api/v1/photos": {
			"get": {
				"operationId": "listPhotos",
				"parameters": [
					{
						"name": "filterRaw",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["all", "raw", "standard"],
							"default": "all"
						}
					},
					{
						"name": "folder",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "camera",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "lens",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "iso",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "dateMonth",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
						}
					},
					{
						"name": "minRating",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 5
						}
					},
					{
						"name": "flag",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["pick", "reject", "unflagged"]
						}
					}
				],
				"responses": {
					"200": {
						"description": "All matching photos",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PhotosResponse"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/InvalidRequest"
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		},
		"/api/v1/photos/{id}": {
			"get": {
				"operationId": "getPhoto",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					}
				],
				"responses": {
					"200": {
						"description": "Photo detail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Photo"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/InvalidRequest"
					},
					"404": {
						"description": "Photo not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"photoNotFound": {
										"value": {
											"error": {
												"code": "PHOTO_NOT_FOUND",
												"message": "Photo not found"
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			},
			"patch": {
				"operationId": "updatePhotoCuration",
				"description": "Sets the photo's star rating and/or pick/reject flag. `flag: null` clears the flag. Available regardless of the native scan mutation flag.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PhotoCurationPatch"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Updated photo detail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Photo"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/InvalidRequest"
					},
					"404": {
						"description": "Photo not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"photoNotFound": {
										"value": {
											"error": {
												"code": "PHOTO_NOT_FOUND",
												"message": "Photo not found"
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		},
		"/api/v1/photos/{id}/similar": {
			"get": {
				"operationId": "listSimilarPhotos",
				"description": "Photos nearest to the source photo by CLIP vector distance (L2), nearest first with ties ordered by ascending id. The source photo is excluded. Only completed vectors for the current model and committed thumbnail generation participate. Optional `minRating` and `flag` filters apply to candidates before ranking and the limit, never to the source. When the source photo has no usable vector, `indexed` is false and `photos` is empty.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 100,
							"default": 30
						}
					},
					{
						"name": "minRating",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 5
						}
					},
					{
						"name": "flag",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["pick", "reject", "unflagged"]
						}
					}
				],
				"responses": {
					"200": {
						"description": "Ranked visually similar photos",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SimilarPhotosResponse"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/InvalidRequest"
					},
					"404": {
						"description": "Photo not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"photoNotFound": {
										"value": {
											"error": {
												"code": "PHOTO_NOT_FOUND",
												"message": "Photo not found"
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		},
		"/api/v1/search": {
			"post": {
				"operationId": "searchPhotos",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SearchRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Ranked photo search results",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SearchResponse"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/InvalidRequest"
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		},
		"/api/v1/scans": {
			"post": {
				"operationId": "startScan",
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StartScanRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Scan start domain result",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StartScanResponse"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/InvalidRequest"
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					},
					"503": {
						"description": "Native scan mutations are disabled",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"nativeScanDisabled": {
										"value": {
											"error": {
												"code": "NATIVE_SCAN_DISABLED",
												"message": "Native scan mutations are disabled"
											}
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/scans/active": {
			"get": {
				"operationId": "listActiveScans",
				"responses": {
					"200": {
						"description": "Active scans in deterministic recovery order",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ActiveScansResponse"
								}
							}
						}
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		},
		"/api/v1/scans/{jobId}": {
			"get": {
				"operationId": "getScan",
				"parameters": [
					{
						"name": "jobId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Durable scan status or null when the UUID is unknown",
						"content": {
							"application/json": {
								"schema": {
									"oneOf": [
										{
											"$ref": "#/components/schemas/Scan"
										},
										{
											"type": "null"
										}
									]
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/InvalidRequest"
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Folder": {
				"type": "object",
				"additionalProperties": false,
				"required": ["name", "path", "photoCount", "children"],
				"properties": {
					"name": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"photoCount": {
						"type": "integer",
						"minimum": 0
					},
					"children": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Folder"
						}
					}
				}
			},
			"FoldersResponse": {
				"type": "object",
				"additionalProperties": false,
				"required": ["folders", "totalPhotos"],
				"properties": {
					"folders": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Folder"
						}
					},
					"totalPhotos": {
						"type": "integer",
						"minimum": 0
					}
				}
			},
			"FilterOptionsResponse": {
				"type": "object",
				"additionalProperties": false,
				"required": ["cameras", "lenses", "isos", "dates"],
				"properties": {
					"cameras": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"lenses": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"isos": {
						"type": "array",
						"items": {
							"type": "integer"
						}
					},
					"dates": {
						"type": "array",
						"items": {
							"type": "string",
							"pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
						}
					}
				}
			},
			"PhotoExif": {
				"type": "object",
				"additionalProperties": false,
				"required": [
					"id",
					"photoId",
					"cameraMake",
					"cameraModel",
					"lensMake",
					"lensModel",
					"focalLength",
					"iso",
					"aperture",
					"shutterSpeed",
					"exposureBias",
					"dateTaken",
					"gpsLatitude",
					"gpsLongitude",
					"gpsAltitude"
				],
				"properties": {
					"id": {
						"type": "integer",
						"minimum": 1
					},
					"photoId": {
						"type": "integer",
						"minimum": 1
					},
					"cameraMake": {
						"type": ["string", "null"]
					},
					"cameraModel": {
						"type": ["string", "null"]
					},
					"lensMake": {
						"type": ["string", "null"]
					},
					"lensModel": {
						"type": ["string", "null"]
					},
					"focalLength": {
						"type": ["integer", "null"]
					},
					"iso": {
						"type": ["integer", "null"]
					},
					"aperture": {
						"type": ["string", "null"]
					},
					"shutterSpeed": {
						"type": ["string", "null"]
					},
					"exposureBias": {
						"type": ["string", "null"]
					},
					"dateTaken": {
						"type": ["string", "null"]
					},
					"gpsLatitude": {
						"type": ["string", "null"]
					},
					"gpsLongitude": {
						"type": ["string", "null"]
					},
					"gpsAltitude": {
						"type": ["string", "null"]
					}
				}
			},
			"Photo": {
				"type": "object",
				"additionalProperties": false,
				"required": [
					"id",
					"path",
					"name",
					"size",
					"createdAt",
					"modifiedAt",
					"width",
					"height",
					"mimeType",
					"isRaw",
					"rawFormat",
					"rawStatus",
					"rawError",
					"thumbnailStatus",
					"thumbnailUpdatedAt",
					"embeddingStatus",
					"phashStatus",
					"rating",
					"flag",
					"exif"
				],
				"properties": {
					"id": {
						"type": "integer",
						"minimum": 1
					},
					"path": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"size": {
						"type": "integer",
						"minimum": 0
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"modifiedAt": {
						"type": "string",
						"format": "date-time"
					},
					"width": {
						"type": ["integer", "null"]
					},
					"height": {
						"type": ["integer", "null"]
					},
					"mimeType": {
						"type": ["string", "null"]
					},
					"isRaw": {
						"type": ["boolean", "null"]
					},
					"rawFormat": {
						"type": ["string", "null"]
					},
					"rawStatus": {
						"type": ["string", "null"]
					},
					"rawError": {
						"type": ["string", "null"]
					},
					"thumbnailStatus": {
						"type": ["string", "null"]
					},
					"thumbnailUpdatedAt": {
						"oneOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						]
					},
					"embeddingStatus": {
						"type": ["string", "null"]
					},
					"phashStatus": {
						"type": ["string", "null"]
					},
					"rating": {
						"type": "integer",
						"minimum": 0,
						"maximum": 5,
						"description": "Star rating; 0 means unrated"
					},
					"flag": {
						"type": ["string", "null"],
						"enum": ["pick", "reject", null]
					},
					"exif": {
						"oneOf": [
							{
								"$ref": "#/components/schemas/PhotoExif"
							},
							{
								"type": "null"
							}
						]
					}
				}
			},
			"PhotoCurationPatch": {
				"type": "object",
				"additionalProperties": false,
				"minProperties": 1,
				"description": "At least one of `rating` or `flag` is required.",
				"properties": {
					"rating": {
						"type": "integer",
						"minimum": 0,
						"maximum": 5
					},
					"flag": {
						"type": ["string", "null"],
						"enum": ["pick", "reject", null],
						"description": "`null` clears the flag"
					}
				}
			},
			"PhotosResponse": {
				"type": "object",
				"additionalProperties": false,
				"required": ["photos", "total", "rawCount"],
				"properties": {
					"photos": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Photo"
						}
					},
					"total": {
						"type": "integer",
						"minimum": 0
					},
					"rawCount": {
						"type": "integer",
						"minimum": 0
					}
				}
			},
			"SearchRequest": {
				"type": "object",
				"additionalProperties": false,
				"required": ["query"],
				"properties": {
					"query": {
						"type": "string",
						"minLength": 1
					},
					"limit": {
						"type": "integer",
						"minimum": 1,
						"maximum": 100,
						"default": 20
					},
					"filterRaw": {
						"type": "string",
						"enum": ["all", "raw", "standard"],
						"default": "all"
					},
					"folder": {
						"type": "string",
						"description": "Restricts results to direct children of this folder"
					},
					"camera": {
						"type": "string"
					},
					"lens": {
						"type": "string"
					},
					"iso": {
						"type": "integer"
					},
					"dateMonth": {
						"type": "string",
						"pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
					},
					"minRating": {
						"type": "integer",
						"minimum": 1,
						"maximum": 5,
						"description": "Only photos rated at least this many stars"
					},
					"flag": {
						"type": "string",
						"enum": ["pick", "reject", "unflagged"],
						"description": "Only picks, only rejects, or only photos without a flag"
					}
				}
			},
			"SearchResponse": {
				"type": "object",
				"additionalProperties": false,
				"required": ["photos", "total", "query"],
				"properties": {
					"photos": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Photo"
						}
					},
					"total": {
						"type": "integer",
						"minimum": 0
					},
					"query": {
						"type": "string"
					}
				}
			},
			"SimilarPhotosResponse": {
				"type": "object",
				"additionalProperties": false,
				"required": ["photos", "total", "sourcePhotoId", "indexed"],
				"properties": {
					"photos": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Photo"
						}
					},
					"total": {
						"type": "integer",
						"minimum": 0
					},
					"sourcePhotoId": {
						"type": "integer",
						"minimum": 1
					},
					"indexed": {
						"type": "boolean"
					}
				}
			},
			"StartScanRequest": {
				"type": "object",
				"additionalProperties": false,
				"properties": {
					"force": {
						"type": "boolean",
						"default": false
					}
				}
			},
			"StartScanSuccess": {
				"type": "object",
				"additionalProperties": false,
				"required": ["success", "jobId"],
				"properties": {
					"success": {
						"const": true
					},
					"jobId": {
						"type": "string",
						"format": "uuid"
					}
				}
			},
			"StartScanFailure": {
				"type": "object",
				"additionalProperties": false,
				"required": ["success", "error"],
				"properties": {
					"success": {
						"const": false
					},
					"error": {
						"const": "The scan could not be started"
					},
					"jobId": {
						"type": "string",
						"format": "uuid"
					}
				}
			},
			"StartScanResponse": {
				"oneOf": [
					{
						"$ref": "#/components/schemas/StartScanSuccess"
					},
					{
						"$ref": "#/components/schemas/StartScanFailure"
					}
				]
			},
			"Scan": {
				"type": "object",
				"additionalProperties": false,
				"required": [
					"id",
					"phase",
					"current",
					"total",
					"status",
					"error",
					"createdAt",
					"updatedAt"
				],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"phase": {
						"type": "string",
						"enum": [
							"queued",
							"discovering",
							"processing",
							"scan-complete",
							"embedding",
							"completed",
							"failed"
						]
					},
					"current": {
						"type": "integer",
						"minimum": 0
					},
					"total": {
						"type": "integer",
						"minimum": 0
					},
					"status": {
						"type": "string",
						"enum": ["queued", "running", "completed", "failed"]
					},
					"error": {
						"description": "Null or a stable public-safe failure message",
						"type": ["string", "null"]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"ActiveScansResponse": {
				"type": "object",
				"additionalProperties": false,
				"required": ["jobs"],
				"properties": {
					"jobs": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Scan"
						}
					}
				}
			},
			"ErrorResponse": {
				"type": "object",
				"additionalProperties": false,
				"required": ["error"],
				"properties": {
					"error": {
						"type": "object",
						"additionalProperties": false,
						"required": ["code", "message"],
						"properties": {
							"code": {
								"type": "string"
							},
							"message": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"responses": {
			"InvalidRequest": {
				"description": "Invalid request",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorResponse"
						},
						"examples": {
							"invalidRequest": {
								"value": {
									"error": {
										"code": "INVALID_REQUEST",
										"message": "Request validation failed"
									}
								}
							}
						}
					}
				}
			},
			"InternalError": {
				"description": "Internal server error",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorResponse"
						},
						"examples": {
							"internalError": {
								"value": {
									"error": {
										"code": "INTERNAL_ERROR",
										"message": "The request could not be completed"
									}
								}
							}
						}
					}
				}
			}
		}
	}
}